refactor(osio): rename wit package to space
